move insertButton to strict registration
