+++ src/contentScript/messenger/strict/api.ts
@@ -63,3 +63,4 @@ export const getAttributeExamples = getMethod("GET_ATTRIBUTE_EXAMPLES");
 export const selectElement = getMethod("SELECT_ELEMENT");
 
 export const insertPanel = getMethod("INSERT_PANEL");
+export const insertButton = getMethod("INSERT_BUTTON");
