+++ src/pageEditor/starterBricks/panel.ts
@@ -37,7 +37,7 @@ import { getDomain } from "@/permissions/patterns";
 import { faWindowMaximize } from "@fortawesome/free-solid-svg-icons";
 import { type ElementConfig } from "@/pageEditor/starterBricks/elementConfig";
 import PanelConfiguration from "@/pageEditor/tabs/panel/PanelConfiguration";
-import { insertPanel } from "@/contentScript/messenger/api";
+import { insertPanel } from "@/contentScript/messenger/strict/api";
 import {
   type DynamicDefinition,
   type PanelSelectionResult,
